requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
biopython>=1.82
httpx>=0.25.0
pydantic>=2.5.0
python-multipart>=0.0.6
